refactor: Dealer에서 View 의존성 제거 및 게임 흐름 제어를 BlackjackGame으로 이동
